Add ability to specify filename for session data
This change allows for the specification of a file name into which the
session data returned from the "login" API call can be saved.
Previously, this data was always saved in a file called "sid.json" which
is still used as the default file name if one is not provided.
Therefore, this is not a breaking change.

This simple change should allow for users who need to manage multiple
Checkpoint resources to do so via standard terraform provider aliases
and/or environment variables.  This was previously impossible due to the
provider overwriting the session data file even when different
endpoints were being contacted.