Refactor complex conditional for increased clarity
Returning the variable 'err' when its value is nil is easy to breeze
over and leaves the reader wondering why this function creates the
session file if it doesn't exist if it is still going to report an error
because it is empty.  Obviously, that isn't what this code does; it
returns an empty session when the session file is empty.  This
refactoring makes that more obvious.